refactory angular app structure
