+++ angular-admin-ui/src/app/app-routing.module.ts
@@ -1,9 +1,9 @@
 import { NgModule }             from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 
-import { ConfigComponent }   from './config/config.component';
-import { DelayedTasksComponent } from './delayed-task/delayed-tasks.component';
-import { DelayedTaskDetailComponent } from './delayed-task/delayed-task-detail.component';
+import { ConfigComponent }   from './components/config/index/config.component';
+import { DelayedTasksComponent } from './components/delayed-task/index/delayed-tasks.component';
+import { DelayedTaskDetailComponent } from './components/delayed-task/show/delayed-task-detail.component';
 
 const routes: Routes = [
   { path: '', redirectTo: '/config', pathMatch: 'full' },
