+++ angular-admin-ui/src/app/components/server/static/angular_app/index.html
@@ -1,0 +1,14 @@
+<!doctype html>
+<html lang="en">
+<head>
+  <meta charset="utf-8">
+  <title>AngularAdminUi</title>
+  <base href="/">
+
+  <meta name="viewport" content="width=device-width, initial-scale=1">
+  <link rel="icon" type="image/x-icon" href="favicon.ico">
+</head>
+<body>
+  <app-root></app-root>
+<script type="text/javascript" src="../../../../../../../server/static/angular_app/inline.bundle.js"></script><script type="text/javascript" src="../../../../../../../server/static/angular_app/polyfills.bundle.js"></script><script type="text/javascript" src="../../../../../../../server/static/angular_app/styles.bundle.js"></script><script type="text/javascript" src="../../../../../../../server/static/angular_app/vendor.bundle.js"></script><script type="text/javascript" src="../../../../../../../server/static/angular_app/main.bundle.js"></script></body>
+</html>
