--- angular-admin-ui/src/app/config/config.service.spec.ts
@@ -1,15 +1,0 @@
-import { TestBed, inject } from '@angular/core/testing';
-
-import { ConfigService } from './config.service';
-
-describe('ConfigService', () => {
-  beforeEach(() => {
-    TestBed.configureTestingModule({
-      providers: [ConfigService]
-    });
-  });
-
-  it('should be created', inject([ConfigService], (service: ConfigService) => {
-    expect(service).toBeTruthy();
-  }));
-});
