+++ angular-admin-ui/src/app/pipes/keys.pipe.ts
